Fix warning
git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder@105389 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libbdk-os/bdk-exception-handler.c
@@ -35,7 +35,7 @@ void __bdk_exception_handler(except_regs_t *exc)
     uint64_t *stack = (uint64_t *)exc->gpr[31];
     for (int i = 0; i < 16; i++)
     {
-        printf("stack[0x%016lx] = 0x%016lx\n", stack, *stack);
+        printf("stack[0x%016lx] = 0x%016lx\n", (uint64_t)stack, *stack);
         stack++;
     }
     printf(BANNER);
